requirements.txt
groq
python-dotenv
pillow
rembg
requests
